Extract ConversationDeclaration helpers
